Export each page content from its parent page
